backend/requirements.txt
fastapi>=0.112,<1
uvicorn[standard]>=0.30,<1
requests>=2.32,<3
beautifulsoup4>=4.12,<5
pdfplumber>=0.11,<1
pydantic>=2.7,<3
python-dotenv>=1.0,<2
PyPDF2>=3.0,<4
PyJWT>=2.9,<3

# Voice + Agents (optional at runtime for LiveKit agent)
livekit-agents[openai,silero,deepgram,cartesia,turn-detector]~=1.2
livekit-plugins-noise-cancellation~=0.2
cerebras-cloud-sdk>=1.2,<2

# CORS if needed later
starlette>=0.37,<1
python-multipart>=0.0.9

